Sort import alphabetically and both post/get methods under the same HttpServlet
